requirements.txt
openai>=1.12.0
python-dotenv==1.1.1
langgraph>=0.2.19
langchain>=0.3.0
langchain-openai>=0.2.0
pyyaml>=6.0
rich>=13.7.0
tavily-python>=0.3.0
langchain-community>=0.3.0
faiss-cpu==1.13.2
numpy>=1.24.0
